Make admin bar cookie persistant
Set a default expiration time for the remote admin bar cookie to match
the expiration time of the WP auth cookie for users who select "Remember
me" option. Otherwise the cookie is only a session cookie and we'll have
to requery logged in state every time the user visits the site.