Remove async/await syntax for ES2015 compatability
The async/await syntax in this module caused issues in environments
without support for generators. While it's possible to fix this by
changing the babel target in this repo, it's just as easy, and probably
easier to read, to just return a promise rather than an awaited result -
the end use is the same in either case.