Send CORS headers in response to preflight request
Some browsers, such as Chrome, now send an unauthenticated OPTIONS
request to an XHR endpoint before sending credentials. If the admin-ajax
endpoint doesn't send the proper header in response to this preflight
request, the credentials will never be sent.

In practice, this meant that users in Chrome were not seeing the admin
bar in cases where they should have been.